change to ORB
ideally i would modify this so that in 6v6 when you press the respawn bind, you do not go to scout, soldier, demo or medic as opposed to random classes.

This prevents you from changing to someone elses class and stopping them from selecting it, which is a bit of an oversight, even though the situation is rare. but i am too lazy to implement it, so maybe once i get bored failing to get Android 12 to build for 7420 devices I will make it